requirements.txt
alembic
click
Flask
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
greenlet
importlib-metadata
importlib-resources
itsdangerous
Jinja2
Mako
MarkupSafe
python-dateutil
python-dotenv
six
SQLAlchemy
Werkzeug
WTForms
zipp
